Move classes into separate packages
