+++ src/Simulation/models/AdHocCar.java
@@ -1,4 +1,4 @@
-package simulation;
+package simulation.models;
 
 import java.util.Random;
 import java.awt.*;
